+++ stockMate/src/main/webapp/WEB-INF/views/dashboard.jsp
@@ -8,6 +8,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">    
     <title>대시보드</title>
     <style>
+    
         /* Reset and Global Styles */
         body, html {
             margin: 0;
